+++ src/Files.App/Data/Models/NavigationInteractionTracker.cs
@@ -10,7 +10,7 @@ using Microsoft.UI.Xaml.Input;
 using System.Diagnostics.CodeAnalysis;
 using System.Numerics;
 
-namespace Files.App.Views
+namespace Files.App.Data.Models
 {
 	internal class NavigationInteractionTracker : IDisposable
 	{
@@ -214,7 +214,7 @@ namespace Files.App.Views
 					}
 				}
 				else
-				{					
+				{
 					_parent._tracker.TryUpdatePositionWithAnimation(_returnAnimation);
 				}
 				_shouldBounceBack = false;
