+++ src/Files.App/UserControls/FilePreviews/CodePreview.xaml.cs
@@ -1,5 +1,5 @@
 using ColorCode;
-using Files.App.ViewModels.Previews;
+using Files.App.ViewModels.UserControls.Previews;
 using Microsoft.UI.Xaml.Controls;
 using System;
