+++ src/Files.App/UserControls/Widgets/FileTagsWidget.xaml.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See the LICENSE.
 
 using Files.App.Helpers.ContextFlyouts;
+using Files.App.ViewModels.UserControls.Widgets;
 using Files.App.ViewModels.Widgets;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
