+++ src/Files.App/UserControls/Widgets/WidgetsListControl.xaml.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.ViewModels.Widgets;
+using Files.App.ViewModels.UserControls.Widgets;
 using Microsoft.UI.Xaml.Controls;
 using System;
