+++ src/Files.App/ViewModels/ContentLayouts/BaseLayoutViewModel.cs
