+++ src/Files.App/ViewModels/UserControls/Previews/BasicPreviewViewModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.ViewModels.Previews
+namespace Files.App.ViewModels.UserControls.Previews
 {
 	public class BasicPreviewViewModel : BasePreviewModel
 	{
