+++ src/Files.App/ViewModels/UserControls/Previews/HtmlPreviewViewModel.cs
@@ -3,7 +3,7 @@
 
 using Files.App.ViewModels.Properties;
 
-namespace Files.App.ViewModels.Previews
+namespace Files.App.ViewModels.UserControls.Previews
 {
 	public class HtmlPreviewViewModel : BasePreviewModel
 	{
