+++ src/Files.App/ViewModels/UserControls/Previews/TextPreviewViewModel.cs
@@ -4,7 +4,7 @@
 using Files.App.UserControls.FilePreviews;
 using Files.App.ViewModels.Properties;
 
-namespace Files.App.ViewModels.Previews
+namespace Files.App.ViewModels.UserControls.Previews
 {
 	public class TextPreviewViewModel : BasePreviewModel
 	{
