+++ src/Files.App/ViewModels/UserControls/Widgets/FileTagsItemViewModel.cs
