+++ src/Files.App/ViewModels/UserControls/Widgets/FileTagsWidgetViewModel.cs
@@ -1,10 +1,11 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
+using Files.App.ViewModels.Widgets;
 using Files.Core.Services;
 using Files.Shared.Utils;
 
-namespace Files.App.ViewModels.Widgets
+namespace Files.App.ViewModels.UserControls.Widgets
 {
 	public sealed partial class FileTagsWidgetViewModel : ObservableObject, IAsyncInitialize
 	{
