+++ src/Files.App/ViewModels/UserControls/Widgets/ICustomWidgetItemModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.ViewModels.Widgets
+namespace Files.App.ViewModels.UserControls.Widgets
 {
 	/// <summary>
 	/// This interface is used to mark widgets that are not standard to Files i.e. custom widgets
