+++ src/Files.App/Views/ContentLayouts/BaseLayout.cs
