+++ src/Files.App/Views/ContentLayouts/ColumnViewBase.xaml
