+++ src/Files.App/Views/ContentLayouts/ColumnViewBrowser.xaml
