+++ src/Files.App/Views/ContentLayouts/DetailsLayoutBrowser.xaml
