+++ src/Files.App/Views/ContentLayouts/GridViewBrowser.xaml.cs
