+++ src/Files.App/Views/ContentLayouts/IBaseLayout.cs
