+++ src/Files.App/Views/ContentLayouts/StandardLayoutMode.cs
